Refactor : 오늘의 토픽룸 API
