+++ STORIX-Api/src/main/java/com/storix/api/domain/topicroom/usecase/TopicRoomUseCaseV2.java
@@ -16,8 +16,11 @@ public class TopicRoomUseCaseV2 {
     private final TopicRoomService topicRoomService;
 
     public Slice<TopicRoomResponseDto> getMyJoinedRooms(Long userId, Pageable pageable) {
-
         return topicRoomService.getMyJoinedRooms(userId, pageable);
+    }
+
+    public List<TopicRoomResponseDto> getTodayTopicRooms(Long userId) {
+        return topicRoomService.getTodayTopicRooms(userId);
     }
 
 }
